#1 - Exclude Spring Data REST test
This test is not related to this project's business logic test. It is just data rest function and it is too heavy to run since it is integration test also influencing to DB. But since it's reasonable to learn it for study purpose it will be excluded not be deleted.