#16- Implementation of user account domain and reflection of erd
Create member account domain with `UserAccount` name

Avoid `user`, `account`, etc. as they are mysql reserved words

The field name is also determined by being conscious of mysql reserved words.

Register 1 random account for test data
It's for testing, but it's a method that exposes the password, so need to think about changing later on 

Member information should not be exposed through API.
To this add a test to check for 404 not found
Easily respond to all http methods